adding session_list_item.xml and recyclerview in tab_sessions.xml. Moving files
