--- app/src/main/java/com/example/bookstats/features/library/tabs/GeneralTab.kt
@@ -1,22 +1,0 @@
-package com.example.bookstats.features.library.tabs
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.bookstats.R
-
-class GeneralTab : Fragment() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.tab_general, container, false)
-    }
-}
