--- app/src/main/java/com/example/bookstats/features/library/tabs/SessionsTab.kt
@@ -1,23 +1,0 @@
-package com.example.bookstats.features.library.tabs
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.bookstats.R
-
-class SessionsTab : Fragment() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.tab_sessions, container, false)
-    }
-
-}
